+++ tests/TestCase/Config/GlobalFunctionsTest.php
@@ -29,7 +29,7 @@ use Cake\TestSuite\TestCase;
 /**
  * GlobalFunctionsTest class.
  *
- * It tests global functions declared in `config/global_functions.php`.
+ * It tests global functions declared in `config/functions/global.php`.
  */
 class GlobalFunctionsTest extends TestCase
 {
@@ -78,54 +78,6 @@ class GlobalFunctionsTest extends TestCase
         
         $result = am($firstTestArray, $secondTestArray, $testString);
         $expected = array_merge($firstTestArray, $secondTestArray, (array)$testString);
-        $this->assertEquals($expected, $result);
-    }
-    
-    /**
-     * Test for `buttonClass()` global function
-     * @return void
-     * @test
-     */
-    public function testButtonClass()
-    {
-        $result = buttonClass();
-        $expected = ['class' => 'btn btn-default'];
-        $this->assertEquals($expected, $result);
-        
-        $result = buttonClass(['class' => 'my-class']);
-        $expected = ['class' => 'my-class btn btn-default'];
-        $this->assertEquals($expected, $result);
-        
-        $result = buttonClass(['class' => 'btn']);
-        $expected = ['class' => 'btn btn-default'];
-        $this->assertEquals($expected, $result);
-        
-        $result = buttonClass(['class' => 'btn-primary']);
-        $expected = ['class' => 'btn-primary btn'];
-        $this->assertEquals($expected, $result);
-        
-        $result = buttonClass(['class' => 'btn btn-primary']);
-        $expected = ['class' => 'btn btn-primary'];
-        $this->assertEquals($expected, $result);
-        
-        $result = buttonClass(['class' => 'my-class btn']);
-        $expected = ['class' => 'my-class btn btn-default'];
-        $this->assertEquals($expected, $result);
-        
-        $result = buttonClass(['class' => 'my-class btn'], 'primary');
-        $expected = ['class' => 'my-class btn btn-primary'];
-        $this->assertEquals($expected, $result);
-        
-        $result = buttonClass(['class' => 'my-class btn-primary']);
-        $expected = ['class' => 'my-class btn-primary btn'];
-        $this->assertEquals($expected, $result);
-        
-        $result = buttonClass(['class' => 'my-class btn btn-primary']);
-        $expected = ['class' => 'my-class btn btn-primary'];
-        $this->assertEquals($expected, $result);
-        
-        $result = buttonClass([], 'primary');
-        $expected = ['class' => 'btn btn-primary'];
         $this->assertEquals($expected, $result);
     }
     
@@ -285,147 +237,6 @@ class GlobalFunctionsTest extends TestCase
         $this->assertFalse(isUrl(DS . 'folder'));
         $this->assertFalse(isUrl(DS . 'folder' . DS));
         $this->assertFalse(isUrl(DS . 'folder' . DS . 'file.txt'));
-    }
-    
-    /**
-     * Test for `optionDefaults()` global function
-     * @return void
-     * @test
-     */
-    public function testOptionDefault()
-    {
-        $options = ['value1' => 'val-1'];
-        
-        $result = optionDefaults(['class' => 'my-class'], $options);
-        $expected = [
-            'value1' => 'val-1',
-            'class' => 'my-class',
-        ];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionDefaults(['class' => 'first-class second-class'], $options);
-        $expected = [
-            'value1' => 'val-1',
-            'class' => 'first-class second-class',
-        ];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionDefaults(['class' => ['first-class', 'second-class']], $options);
-        $expected = [
-            'value1' => 'val-1',
-            'class' => 'first-class second-class',
-        ];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionDefaults(['class' => ['first-class', 'second-class', 'first-class']], $options);
-        $expected = [
-            'value1' => 'val-1',
-            'class' => 'first-class second-class',
-        ];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionDefaults(['class' => ['first-class', ['second-class']]], $options);
-        $expected = [
-            'value1' => 'val-1',
-            'class' => 'first-class second-class',
-        ];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionDefaults(['class' => ['first-class', ['second-class', ['third-class']]]], $options);
-        $expected = [
-            'value1' => 'val-1',
-            'class' => 'first-class second-class third-class',
-        ];
-        $this->assertEquals($expected, $result);
-        
-        //This doesn't change the value
-        $result = optionDefaults(['value1' => 'new-val-1'], $options);
-        $expected = ['value1' => 'val-1'];
-        $this->assertEquals($expected, $result);
-        
-        //Backward compatibility with three arguments
-        $result = optionDefaults('class', 'my-class', $options);
-        $expected = [
-            'value1' => 'val-1',
-            'class' => 'my-class',
-        ];
-        $this->assertEquals($expected, $result);
-        
-        //Backward compatibility with three arguments
-        $result = optionDefaults('class', 'first-class second-class', $options);
-        $expected = [
-            'value1' => 'val-1',
-            'class' => 'first-class second-class',
-        ];
-        $this->assertEquals($expected, $result);
-    }
-    
-    /**
-     * Test for `optionValues()` global function
-     * @return void
-     * @test
-     */
-    public function testOptionValue()
-    {
-        $options = ['value1' => 'alfa beta'];
-        
-        $result = optionValues(['class' => 'my-class'], $options);
-        $expected = ['value1' => 'alfa beta', 'class' => 'my-class'];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionValues(['class' => ['my-class']], $options);
-        $expected = ['value1' => 'alfa beta', 'class' => 'my-class'];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionValues(['value1' => 'beta'], $options);
-        $expected = ['value1' => 'alfa beta'];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionValues(['value1' => 'gamma'], $options);
-        $expected = ['value1' => 'alfa beta gamma'];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionValues(['value1' => 'delta gamma'], $options);
-        $expected = ['value1' => 'alfa beta delta gamma'];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionValues(['value1' => 'delta gamma delta'], $options);
-        $expected = ['value1' => 'alfa beta delta gamma'];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionValues(['value1' => ['delta', 'gamma']], $options);
-        $expected = ['value1' => 'alfa beta delta gamma'];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionValues(['value1' => ['delta', ['gamma']]], $options);
-        $expected = ['value1' => 'alfa beta delta gamma'];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionValues(['value1' => ['delta', ['gamma', ['ypsilon']]]], $options);
-        $expected = ['value1' => 'alfa beta delta gamma ypsilon'];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionValues([
-            'class' => 'my-class',
-            'value1' => 'gamma'
-        ], $options);
-        $expected = [
-            'class' => 'my-class',
-            'value1' => 'alfa beta gamma',
-        ];
-        $this->assertEquals($expected, $result);
-        
-        $result = optionValues([
-            'class' => 'my-class',
-            'value1' => ['delta', 'gamma']
-        ], $options);
-        $expected = ['class' => 'my-class', 'value1' => 'alfa beta delta gamma'];
-        $this->assertEquals($expected, $result);
-
-        //Backward compatibility with three arguments
-        $result = optionValues('value1', 'gamma', $options);
-        $expected = ['value1' => 'alfa beta gamma'];
-        $this->assertEquals($expected, $result);
     }
     
     /**